Fix addin installer on pure Win11 systems.
The "WOW6432Node" segment is automatically (and always) injected on Windows 11 systems, no need to expilcitly add it here.
So in the final path, the segments exists twice and EA does not find the registered DLL.

On Win10 systems, it's not automatically injected, but also not needed. It works without.